requirements.txt
Django==3.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.12.1
isort==5.6.4
black==20.8b1
flake8==3.8.4
pre-commit==2.9.3
pytest==6.2.1
django-allauth==0.44.0
dj-rest-auth==2.1.3
django-rest-knox==4.1.0
